pyproject.toml
[tool.poetry]
name = "yebigun"
version = "0.1.0"
description = ""
authors = ["Damien-IR <qq23667@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
playwright = "^1.28.0"
PyYAML = "^6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
